requirements.txt
streamlit==1.29.0
requests==2.31.0
pillow==9.4.0
numpy==1.26.2
matplotlib==3.8.2
scikit-learn==1.3.2